Update Crazy.txt (#26)
So there are small alterations - in first parts od verses 2 & 3. I based this change on this version https://tabs.ultimate-guitar.com/tab/gnarls-barkley/crazy-chords-377909 the F and Cmaj7 switch is absolutely necessary. When it comes to Am-A it is hard to say. the way i see it its like this -A lifts the melody and gives it right direction, Am is taking it down (it is hard to sing knowing how the orginal goes(at least for me)), which is ok in the second part of the verse.